refactor: calendarService 리팩토링
